这是第一版能完赛的代码
添加远程停止功能，修改编码类型，优化编码器数据类型，更新主函数结构，修复编译日志时间

+++ Project/CODE/UserLibraries/encoder.c
@@ -20,7 +20,7 @@
 
 #include "headfile.h"
 
-static uint16 TravelDistanceL = 0, TravelDistanceR = 0, TravelDistance = 0;
+static uint32 TravelDistanceL = 0, TravelDistanceR = 0, TravelDistance = 0;
 
 /**
  * @brief ³õÊ¼»¯±àÂëÆ÷
